+++ tests/DatabaseTest/CreateFeedDatabase.php
@@ -14,13 +14,13 @@ class CreateFeedDatabase
             $table->increments('id');
             $table->string('title');
             $table->integer('category')->unsigned()->index('feed_category_foreign');
-            $table->text('short');
-            $table->text('description');
+            $table->text('short')->nullable();
+            $table->text('description')->nullable();
             $table->string('url', 191)->unique();
             $table->dateTime('date');
             $table->integer('position')->default(0);
             $table->integer('active')->default(1);
-            $table->integer('user_id')->unsigned()->index('feed_user_id_foreign');
+            $table->integer('user_id')->unsigned()->index('feed_user_id_foreign')->nullable();
             $table->timestamps();
 
             $table->index(['url', 'active', 'category']);